thesis page redesign
