--- src/app/student/thesis/layout.tsx
@@ -1,22 +1,0 @@
-"use client";
-
-import Sidenav from "@/components/organisms/Sidenav";
-import { withAuth } from "@/utilities/AuthWrapper";
-import { Toaster } from "react-hot-toast";
-
-function DashboardLayout({
-    children,
-}: Readonly<{
-    children: React.ReactNode;
-}>) {
-    return (
-        <>
-            <Sidenav>
-                {children}
-            </Sidenav>
-            <Toaster />
-        </>
-    );
-}
-
-export default withAuth(DashboardLayout, ["student"]);
